chore(switch): convert to Functional Component (#1200)
Convert `hv-switch` to Functional Component. Confirm no degradation in
render count.


[Asana](https://app.asana.com/1/47184964732898/project/1204008699308084/task/1210601789543020?focus=true)
